+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Master (unreleased)
 
-- Add `Command#run` method to execute a task on current thread [#13](https://github.com/cookpad/expeditor/pull/13)
+- Add a `current_thread` option of `Expeditor::Command#start` method to execute a task on current thread [#13](https://github.com/cookpad/expeditor/pull/13)
 - Drop support for MRI 2.0.x [#15](https://github.com/cookpad/expeditor/pull/15)
 - Deprecate Expeditor::Command#with_fallback. Use `set_fallback` instead [#14](https://github.com/cookpad/expeditor/pull/14)
 - Do not allow set_fallback call after command is started. [#18](https://github.com/cookpad/expeditor/pull/18)